minor fixes

+++ Capstone/Frontend/Capstone/src/app/posts/posts.component.ts
@@ -1,15 +1,19 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { map } from 'rxjs/operators';
 import Swal from 'sweetalert2';
 import { AuthResponse } from '../models/auth-response';
+import { Comment } from 'src/app/models/comment';
+import { CommentDto } from '../models/commentDto';
 import { Post } from '../models/post';
 import { PostDto } from '../models/postDto';
 import { User } from '../models/user';
 import { AuthService } from '../services/auth.service';
+import { CommentService } from '../services/comment.service';
 import { ImageProcessingService } from '../services/image-processing.service';
 import { PostService } from '../services/post.service';
 import { UserService } from '../services/user.service';
+import { PostCommentsComponent } from './post-comments/post-comments.component';
 
 @Component({
   selector: 'app-posts',
@@ -23,15 +27,16 @@ export class PostsComponent implements OnInit {
   formAction: string = 'create'
   userList: User[] = [];
   showForm: boolean = false;
+  posts: Post[] = [];
+
 
   constructor(
     private postSvc: PostService,
     private authSvc: AuthService,
     private userSvc: UserService,
+    private commentSvc: CommentService,
     private imgSvc: ImageProcessingService,
   ) { }
-
-  posts: Post[] = [];
 
   checkLog(): boolean {
     return this.authSvc.isUserLogged()
@@ -66,23 +71,44 @@ export class PostsComponent implements OnInit {
 
   }
 
+  //-------------- POSTS METHODS --------------------
+
   deletePost(post: Post): void {
-    this.postSvc.deletePost(post).subscribe(
-      {
-        next: res => {
-          let index: number = this.posts.findIndex(p => p.id === post.id)
-          this.posts.splice(index, 1);
-          Swal.fire({
-            position: 'top-end',
-            icon: 'success',
-            title: 'Post Deleated',
-            showConfirmButton: false,
-            timer: 2000
-          })
-        },
-        error: error => console.log(error)
+
+    Swal.fire({
+      title: 'warning!',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      cancelButtonText: 'Cancel',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.postSvc.deletePost(post).subscribe(
+          {
+            next: res => {
+              let index: number = this.posts.findIndex(p => p.id === post.id)
+              this.posts.splice(index, 1);
+              Swal.fire({
+                position: 'top-end',
+                icon: 'success',
+                title: 'Post Deleated',
+                showConfirmButton: false,
+                timer: 2000
+              })
+            },
+            error: error => console.log(error)
+          }
+        )
+      } else {
+        Swal.fire({
+          text: 'Post not deleted',
+          icon: 'error',
+          confirmButtonText: 'Ok'
+        })
       }
-    )
+    });
+
   }
 
   addPost(post: PostDto): void {
@@ -93,6 +119,7 @@ export class PostsComponent implements OnInit {
     ).subscribe(res => {
       this.posts.push(res as Post)
       this.currentPost = new Post()
+      this.showForm = false;
       Swal.fire({
         position: 'top-end',
         icon: 'success',
@@ -127,26 +154,66 @@ export class PostsComponent implements OnInit {
   prepareFormDate(post: PostDto): FormData {
 
     const formData = new FormData();
-
     formData.append(
       'post',
       new Blob([JSON.stringify(post)], { type: 'application/json' })
     );
-
     if (post.image)
       formData.append(
         'imageFile',
         post.image,
         post.image?.name
       )
-
     return formData
-
   }
 
   toggleCreatePost() {
     this.formAction = 'create'
     this.showForm = !this.showForm;
+  }
+
+  likePost(postId: number) {
+    this.postSvc.likePost(postId + "", this.loggedUser.id).subscribe(
+      res => {
+        let index = this.posts.findIndex((p: Post) => p.id == res.id)
+        this.posts[index].likes = res.likes
+      }
+
+    )
+  }
+
+  //----------- COMMENTS METHODS ----------------------------
+
+  showComments: boolean = false;
+  newComment: Comment = new Comment();
+
+  @ViewChild(PostCommentsComponent) child!: PostCommentsComponent;
+
+  toogleOpenComments() {
+    this.showComments = !this.showComments;
+  }
+
+  openComments(post: Post) {
+    this.toogleOpenComments()
+    this.currentPost = post;
+  }
+
+  sendComment(comment: CommentDto) {
+
+    this.commentSvc.addComment(comment)
+      .subscribe(res => {
+
+        if(this.currentPost.id) // per disambiguare se il commento arriva dal child PostComments o no
+          this.child.updateComments(res)
+
+        Swal.fire({
+          position: 'top-end',
+          icon: 'success',
+          title: 'Comment sended',
+          showConfirmButton: false,
+          timer: 2000
+        })
+      })
   }
 
 }
